The now playing intent now covers the additional option of getting weekdays (i.e. "what's playing tomorrow", "what's playing Friday"), rather than having spearate intents.
The showtimes intent now separates the showtimes into 3-D and regular movie times. It also filters out showtimes that are already past the listed sell by date so it won't list showtimes that they've stopped selling tickets for.

Added an intent to check the saved city and state.

Added an intent to get the phone number of a theatre (in-progress).