Fixing up the opening text and reprompt. The location doesn't realy matter, just prompt to set a favorite theatre.
Fixing the ListLocalTheatres intent so that the user can specify the city/state to look for theatres in. Make an api call instead of looking in the saved theatre data. If no city/state is specified, it will look for the city/state saved in the user preferences.

Fixing the SetFavoriteTheatre intent. It must have broken with the change to the api call method the if statement was pointing to the wrong variable.